requirements.txt
pytesseract
openai>=1.63.0
python-dotenv>=1.0.1
Pillow>=10.0.0
PyYAML>=6.0.0
mss>=9.0.0
pyautogui>=0.9.54
comtypes>=1.4.0; platform_system=="Windows"
pyobjc-core>=10.0; platform_system=="Darwin"
pyobjc-framework-Cocoa>=10.0; platform_system=="Darwin"
pyobjc-framework-Quartz>=10.0; platform_system=="Darwin"
pyobjc-framework-ApplicationServices>=10.0; platform_system=="Darwin"
pytest>=7.4.0
numpy>=1.26.4
scikit-image>=0.26.0rc1; python_version>="3.14"
scikit-image>=0.22.0; python_version<"3.14"
